Support bytearray and other buffers for password
This change implements support for passing a bytearray
or other type that implements the buffer interface to the
password parameter of authGSSClientInit. ASCII and UTF8
encoded byte strings are supported. See the test suite for
examples.

This change also fixes a bug in authGSSClientUsername when
the username includes non-ascii characters.

Resolves #6 and #7.