Fix explicit principal support
This change fixes the authGSSClientInit principal
parameter. Previously the only way to specify an explicit
principal was through the user, domain, and password parameters.
Those parameters are now deprecated.

The principal parameter now accepts a unicode string or any
8 bit string type that implements the buffer interface. You can
pass just a user principal (user@realm) or a user principal and
password (user@realm:password). The user principal and password
can be percent encoded if either might include the : character.
See the documentation for more details.

Resolves #11.